Added a new test checking 'removed sync' and renamed file
