+++ lib/assets/javascripts/cartodb3/dataset.js
@@ -29,7 +29,7 @@ var VisTableModel = require('./data/visualization-table-model');
 var DatasetUnlockModalView = require('./dataset/dataset-unlock-modal-view');
 var EditorModel = require('./data/editor-model');
 var UserModel = require('./data/user-model');
-var HeaderView = require('./dataset/dataset-header/header-view');
+var HeaderView = require('./dataset/dataset-header/dataset-header-view');
 var Notifier = require('./components/notifier/notifier');
 var UserGroupFetcher = require('./data/users-group-fetcher');
 var SQLUtils = require('./helpers/sql-utils');
