+++ lib/assets/javascripts/cartodb3/dataset/dataset-header/dataset-header-view.js
@@ -1,5 +1,5 @@
 var CoreView = require('backbone/core-view');
-var template = require('./header.tpl');
+var template = require('./dataset-header.tpl');
 var moment = require('moment');
 var InlineEditorView = require('../../components/inline-editor/inline-editor-view');
 var TipsyTooltipView = require('../../components/tipsy-tooltip-view');
@@ -89,7 +89,7 @@ module.exports = CoreView.extend({
   },
 
   _initBinds: function () {
-    this._syncModel.bind('change:state change:interval', this.render, this);
+    this._syncModel.bind('change:state change:interval destroy', this.render, this);
     this.add_related_model(this._syncModel);
     this._visModel.bind('change:name', this._onChangeName, this);
     this._visModel.bind('change:privacy change:name', this.render, this);
