+++ lib/assets/test/spec/cartodb3/dataset/dataset-header-view.spec.js
@@ -1,12 +1,13 @@
 var _ = require('underscore');
 var ConfigModel = require('../../../../javascripts/cartodb3/data/config-model');
-var HeaderView = require('../../../../javascripts/cartodb3/dataset/dataset-header/header-view');
+var HeaderView = require('../../../../javascripts/cartodb3/dataset/dataset-header/dataset-header-view');
 var UserModel = require('../../../../javascripts/cartodb3/data/user-model');
 var TableModel = require('../../../../javascripts/cartodb3/data/table-model');
 var VisModel = require('../../../../javascripts/cartodb3/data/vis-definition-model');
+var SyncModel = require('../../../../javascripts/cartodb3/data/synchronization-model');
 var QuerySchemaModel = require('../../../../javascripts/cartodb3/data/query-schema-model');
 
-describe('dataset/header-view', function () {
+describe('dataset/dataset-header-view', function () {
   beforeEach(function () {
     var configModel = new ConfigModel({
       base_url: '/u/pepe'
@@ -44,7 +45,12 @@ describe('dataset/header-view', function () {
     });
     spyOn(querySchemaModel, 'fetch');
 
-    var syncModel = jasmine.createSpyObj('syncModel', ['get', 'set', 'isSync', 'bind', 'off', 'isSyncing']);
+    this.syncModel = new SyncModel({}, { configModel: configModel });
+    this.syncModel.sync = function (a, b, opts) {
+      opts && opts.success();
+    };
+    spyOn(this.syncModel, 'isSync');
+
     var router = jasmine.createSpyObj('router', ['navigate']);
     var modals = jasmine.createSpyObj('modals', ['create']);
 
@@ -58,7 +64,7 @@ describe('dataset/header-view', function () {
       userModel: userModel,
       tableModel: tableModel,
       visModel: visModel,
-      syncModel: syncModel
+      syncModel: this.syncModel
     }, {
       configModel: this._configModel
     });
@@ -69,7 +75,7 @@ describe('dataset/header-view', function () {
   describe('render', function () {
     it('should render the inline editor view', function () {
       spyOn(this.view._userModel, 'isInsideOrg').and.returnValue(true);
-      this.view._syncModel.isSync.and.returnValue(false);
+      this.syncModel.isSync.and.returnValue(false);
       this.view._tableModel.isOwner.and.returnValue(true);
       expect(_.size(this.view._subviews)).toBe(2);
       expect(this.view.$('.Inline-editor').length).toBe(1);
@@ -78,7 +84,7 @@ describe('dataset/header-view', function () {
 
     it('should not render the inline editor view but sync table info', function () {
       spyOn(this.view._userModel, 'isInsideOrg').and.returnValue(true);
-      this.view._syncModel.isSync.and.returnValue(true);
+      this.syncModel.isSync.and.returnValue(true);
       this.view._tableModel.isOwner.and.returnValue(true);
       this.view.render();
       expect(_.size(this.view._subviews)).toBe(2); // inline editor, with bagde
@@ -89,7 +95,7 @@ describe('dataset/header-view', function () {
     it('should not render the inline editor or the sync info view or the share info view if owner is different', function () {
       spyOn(this.view._userModel, 'isInsideOrg').and.returnValue(true);
       this.view._tableModel.isOwner.and.returnValue(false);
-      this.view._syncModel.isSync.and.returnValue(true);
+      this.syncModel.isSync.and.returnValue(true);
       this.view.render();
       expect(_.size(this.view._subviews)).toBe(0);
       expect(this.view.$('.Inline-editor').length).toBe(0);
@@ -98,7 +104,7 @@ describe('dataset/header-view', function () {
 
     it('should render privacy dropdown when user doesn\'t belong to an organization', function () {
       spyOn(this.view._userModel, 'isInsideOrg').and.returnValue(false);
-      this.view._syncModel.isSync.and.returnValue(false);
+      this.syncModel.isSync.and.returnValue(false);
       this.view._tableModel.isOwner.and.returnValue(true);
       this.view.render();
       expect(_.size(this.view._subviews)).toBe(2);
@@ -145,6 +151,12 @@ describe('dataset/header-view', function () {
       this.view.render();
       this.view.$('.js-privacy').click();
       expect(this.view._modals.create).not.toHaveBeenCalled();
+    });
+
+    it('should render when synchronization has been destroyed', function () {
+      HeaderView.prototype.render.calls.reset();
+      this.syncModel.destroy();
+      expect(HeaderView.prototype.render).toHaveBeenCalled();
     });
   });
